third_party/blink/web_tests/external/wpt/html/canvas/element/manual/imagebitmap/createImageBitmap-drawImage-expected.txt

This is a testharness.js-based test.
Found 10 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] createImageBitmap from an HTMLVideoElement, and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (5, 5) expected 247 +/- 3 but got 254
[FAIL] createImageBitmap from an HTMLVideoElement scaled down, and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (2.5, 2.5) expected 247 +/- 3 but got 254
[FAIL] createImageBitmap from an HTMLVideoElement scaled up, and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (10, 10) expected 247 +/- 3 but got 254
[FAIL] createImageBitmap from an HTMLVideoElement resized, and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (2.5, 10) expected 247 +/- 3 but got 254
[FAIL] createImageBitmap from an HTMLVideoElement with negative sw/sh, and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (5, 5) expected 247 +/- 3 but got 254
[FAIL] createImageBitmap from an HTMLVideoElement from a data URL, and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (5, 5) expected 247 +/- 3 but got 254
[FAIL] createImageBitmap from an HTMLVideoElement from a data URL scaled down, and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (2.5, 2.5) expected 247 +/- 3 but got 254
[FAIL] createImageBitmap from an HTMLVideoElement from a data URL scaled up, and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (10, 10) expected 247 +/- 3 but got 254
[FAIL] createImageBitmap from an HTMLVideoElement from a data URL resized, and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (2.5, 10) expected 247 +/- 3 but got 254
[FAIL] createImageBitmap from an HTMLVideoElement from a data URL with negative sw/sh, and drawImage on the created ImageBitmap
  assert_approx_equals: Red channel of the pixel at (5, 5) expected 247 +/- 3 but got 254
Harness: the test ran to completion.
